+++ CHANGELOG.md
@@ -89,4 +89,5 @@
 
 ## dev
 
-- Added method to specify the format of address components of an `FormattedParsedAddress`. 
+- Added method to specify the format of address components of an `FormattedParsedAddress`. Formatting can specify the
+  field separator, the field to be capitalized, and the field to be upper case.